// src/hooks/orval/model/teamsRecord.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * PocketBase API
 * API documentation generated from PocketBase schema with collection grouping
 * OpenAPI spec version: 1.0.0
 */
import type { TeamsRecordStatus } from './teamsRecordStatus.ts';

export interface TeamsRecord {
  /** Unique record identifier */
  id?: string;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  tournament: string;
  name: string;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  captain?: string;
  status: TeamsRecordStatus;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  created_by?: string;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  updated_by?: string;
  archived: boolean;
  created?: string;
  updated?: string;
}